# [Metric] Accelerate the calculation of F1