## Description ##
Accelerate the calculation of F1 by removing the for-loop. I find that using GPU will in fact make the code slower :sweat_smile:.

## Checklist ##
### Essentials ###
- [x] Passed code style checking (`make lint`)
- [x] Changes are complete (i.e. I finished coding on this PR)
- [x] All changes have test coverage:
- Unit tests are added for small changes to verify correctness (e.g. adding a new operator)
- Nightly tests are added for complicated/long-running ones (e.g. changing distributed kvstore)
- Build tests will be added for build configuration changes (e.g. adding a new build option with NCCL)
- [x] Code is well-documented: 
- For user-facing API changes, API doc string has been updated. 
- For new C++ functions in header files, their functionalities and arguments are documented. 
- For new examples, README.md is added to explain the what the example does, the source of the dataset, expected performance on test set and reference to the original paper if applicable
- [x] To the my best knowledge, examples are either not affected by this change, or have been fixed to be compatible with this change

### Changes ###
- [x] Accelerate F1


Results of running the script in https://github.com/apache/incubator-mxnet/pull/9705/
Before:
```
Metric         Data-Ctx  Label-Ctx   Data Size   Batch Size     Output Dim     Elapsed Time
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      131072      16             2              0.8075
F1             cpu(0)    gpu(0)      131072      16             2              1.7331
F1             gpu(0)    cpu(0)      131072      16             2              1.8095
F1             gpu(0)    gpu(0)      131072      16             2              2.5608
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      131072      64             2              0.48318
F1             cpu(0)    gpu(0)      131072      64             2              0.68605
F1             gpu(0)    cpu(0)      131072      64             2              0.59764
F1             gpu(0)    gpu(0)      131072      64             2              0.71643
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      131072      256            2              0.33182
F1             cpu(0)    gpu(0)      131072      256            2              0.49648
F1             gpu(0)    cpu(0)      131072      256            2              0.43996
F1             gpu(0)    gpu(0)      131072      256            2              0.47916
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      131072      1024           2              0.30854
F1             cpu(0)    gpu(0)      131072      1024           2              0.3868
F1             gpu(0)    cpu(0)      131072      1024           2              0.33063
F1             gpu(0)    gpu(0)      131072      1024           2              0.33789
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      16384       16             2              0.09547
F1             cpu(0)    gpu(0)      16384       16             2              0.24957
F1             gpu(0)    cpu(0)      16384       16             2              0.325
F1             gpu(0)    gpu(0)      16384       16             2              0.30319
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      16384       64             2              0.0871
F1             cpu(0)    gpu(0)      16384       64             2              0.11098
F1             gpu(0)    cpu(0)      16384       64             2              0.097242
F1             gpu(0)    gpu(0)      16384       64             2              0.1261
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      16384       256            2              0.041346
F1             cpu(0)    gpu(0)      16384       256            2              0.055987
F1             gpu(0)    cpu(0)      16384       256            2              0.054622
F1             gpu(0)    gpu(0)      16384       256            2              0.064322
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      16384       1024           2              0.074404
F1             cpu(0)    gpu(0)      16384       1024           2              0.057779
F1             gpu(0)    cpu(0)      16384       1024           2              0.045733
F1             gpu(0)    gpu(0)      16384       1024           2              0.044807
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      2048        16             2              0.012862
F1             cpu(0)    gpu(0)      2048        16             2              0.031966
F1             gpu(0)    cpu(0)      2048        16             2              0.028063
F1             gpu(0)    gpu(0)      2048        16             2              0.032354
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      2048        64             2              0.0078101
F1             cpu(0)    gpu(0)      2048        64             2              0.017859
F1             gpu(0)    cpu(0)      2048        64             2              0.011364
F1             gpu(0)    gpu(0)      2048        64             2              0.014009
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      2048        256            2              0.0052049
F1             cpu(0)    gpu(0)      2048        256            2              0.0073137
F1             gpu(0)    cpu(0)      2048        256            2              0.0071828
F1             gpu(0)    gpu(0)      2048        256            2              0.0076649
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      2048        1024           2              0.0050001
F1             cpu(0)    gpu(0)      2048        1024           2              0.0089076
F1             gpu(0)    cpu(0)      2048        1024           2              0.011183
F1             gpu(0)    gpu(0)      2048        1024           2              0.0092695
------------------------------------------------------------------------------------------
```
After
```
Metric         Data-Ctx  Label-Ctx   Data Size   Batch Size     Output Dim     Elapsed Time
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      131072      16             2              0.69502
F1             cpu(0)    gpu(0)      131072      16             2              1.8135
F1             gpu(0)    cpu(0)      131072      16             2              1.754
F1             gpu(0)    gpu(0)      131072      16             2              2.4507
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      131072      64             2              0.21756
F1             cpu(0)    gpu(0)      131072      64             2              0.30392
F1             gpu(0)    cpu(0)      131072      64             2              0.48876
F1             gpu(0)    gpu(0)      131072      64             2              0.56596
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      131072      256            2              0.070438
F1             cpu(0)    gpu(0)      131072      256            2              0.10176
F1             gpu(0)    cpu(0)      131072      256            2              0.14616
F1             gpu(0)    gpu(0)      131072      256            2              0.17256
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      131072      1024           2              0.027534
F1             cpu(0)    gpu(0)      131072      1024           2              0.049942
F1             gpu(0)    cpu(0)      131072      1024           2              0.031038
F1             gpu(0)    gpu(0)      131072      1024           2              0.035196
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      16384       16             2              0.11917
F1             cpu(0)    gpu(0)      16384       16             2              0.34919
F1             gpu(0)    cpu(0)      16384       16             2              0.22027
F1             gpu(0)    gpu(0)      16384       16             2              0.42634
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      16384       64             2              0.029119
F1             cpu(0)    gpu(0)      16384       64             2              0.049232
F1             gpu(0)    cpu(0)      16384       64             2              0.04909
F1             gpu(0)    gpu(0)      16384       64             2              0.081102
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      16384       256            2              0.0090246
F1             cpu(0)    gpu(0)      16384       256            2              0.014211
F1             gpu(0)    cpu(0)      16384       256            2              0.012915
F1             gpu(0)    gpu(0)      16384       256            2              0.016437
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      16384       1024           2              0.0036936
F1             cpu(0)    gpu(0)      16384       1024           2              0.011009
F1             gpu(0)    cpu(0)      16384       1024           2              0.0092134
F1             gpu(0)    gpu(0)      16384       1024           2              0.0060875
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      2048        16             2              0.015834
F1             cpu(0)    gpu(0)      2048        16             2              0.026115
F1             gpu(0)    cpu(0)      2048        16             2              0.025691
F1             gpu(0)    gpu(0)      2048        16             2              0.030914
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      2048        64             2              0.0044892
F1             cpu(0)    gpu(0)      2048        64             2              0.0088701
F1             gpu(0)    cpu(0)      2048        64             2              0.0074565
F1             gpu(0)    gpu(0)      2048        64             2              0.0094738
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      2048        256            2              0.0012782
F1             cpu(0)    gpu(0)      2048        256            2              0.0021632
F1             gpu(0)    cpu(0)      2048        256            2              0.002121
F1             gpu(0)    gpu(0)      2048        256            2              0.0024769
------------------------------------------------------------------------------------------
F1             cpu(0)    cpu(0)      2048        1024           2              0.00060058
F1             cpu(0)    gpu(0)      2048        1024           2              0.0007112
F1             gpu(0)    cpu(0)      2048        1024           2              0.00064564
F1             gpu(0)    gpu(0)      2048        1024           2              0.00079536
------------------------------------------------------------------------------------------
```