@szha 2018-02-20
Given that those are typical cases, would it make sense to always use cpu?