@sxjscience 2018-02-20
Yes, I think CPU should be used.