@szha 2018-02-21
There's no wait in between the for loops. That should be addressed before calling the results conclusive.