@sxjscience 2018-02-21
We have `.asscalar()` so it should be okay to not call `nd.waitall()` between the for-loops.