@szha 2018-02-21
Ah, sorry I read the results wrong. Looks like numpy sum is an order of magnitude faster than either nd CPU or GPU. What does it look like if we take into account asnumpy()?