@sxjscience 2018-02-21
Our implementations are much slower than numpy if the inputs are small.